COMMIT FIX
feat: Implementación de autenticación JWT y mejoras generales.

- Añadido endpoint de login (/login) en orders_service para autenticar a los usuarios utilizando JWT.
- Módulo de autenticación creado en auth.py que compara credenciales desde el .env.
- Actualización de app.py para proteger los endpoints sensibles con @jwt_required().
- Actualización de Postman para hacer solicitudes autenticadas utilizando tokens JWT en el header de autorización.
- Eliminación de credenciales hardcodeadas, moviendo ADMIN_USERNAME, ADMIN_PASSWORD y SECRET_KEY a archivos .env.
- Corrección de la configuración de bases de datos en los servicios para utilizar rutas absolutas de SQLite.
- Pequeñas mejoras y ajustes de seguridad en la gestión de credenciales.
