+++ .gitignore
@@ -1,5 +1,6 @@
 # Ignore virtual environments
 .venv/
+.env
 
 # Ignore Python bytecode and cache
 __pycache__/
