+++ orders_service/app.py
@@ -1,23 +1,32 @@
 from flask import Flask, jsonify, request
-from flask_sqlalchemy import SQLAlchemy
-import pika
-import json
+from flask_jwt_extended import JWTManager, get_jwt_identity, jwt_required
+from models import db, Order
+from services import send_to_queue
+from dotenv import load_dotenv
+from auth import generate_token 
+import os
+
+load_dotenv()
 
 app = Flask(__name__)
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///orders.db"
 
-db = SQLAlchemy(app)
+app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("DATABASE_URL")
+app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+app.config["JWT_SECRET_KEY"] = os.getenv("SECRET_KEY")
 
-class Order(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    product_id = db.Column(db.Integer, nullable=False)
-    quantity = db.Column(db.Integer, nullable=False)
+db.init_app(app)
+jwt = JWTManager(app)
 
 with app.app_context():
     db.create_all()
-
+@app.route('/login', methods=['POST'])
+def login():
+    username = request.json.get('username', None)
+    password = request.json.get('password', None)
+    return generate_token(username, password)
 
 @app.route("/orders", methods=["POST"])
+@jwt_required()
 def add_order():
     data = request.json
     new_order = Order(product_id=data["product_id"], 
@@ -30,17 +39,6 @@ def add_order():
     return jsonify({
         "message": "order placed"
     }), 201
-
-def send_to_queue(order):
-    connection = pika.BlockingConnection(pika.ConnectionParameters("localhost"))
-    channel = connection.channel()
-    channel.queue_declare(queue="order_queue")
-    message = {
-        "product_id": order.product_id,
-        "quantity": order.quantity
-    }
-    channel.basic_publish(exchange="", routing_key="order_queue", body=json.dumps(message))
-    connection.close()
 
 if __name__ == "__main__":
     app.run(port=5002)
